requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.0
openai==1.3.0
google-ai-generativelanguage==0.4.0
google-generativeai==0.3.1

python-docx==0.8.11
fpdf==1.7.2
pandas==2.0.3
xlsxwriter==3.1.2
python-pptx==0.6.21